requirements.txt
django-admin
djangorestframework
django-filter